chore: extract snyk vulnerability struct
